#7 security (#16)
* Implemented security using JWT tokens

* Added endpoints for registration and login (getting a token by username and password)

* Added roles (admin and user)

* Changed controllers, services and tests according to user rights (for example - a user cannot delete another user's album)

+++ src/main/java/com/example/photobook/service/impl/PhotoServiceImpl.java
@@ -3,9 +3,12 @@ package com.example.photobook.service.impl;
 import com.example.photobook.dto.PhotoDto;
 import com.example.photobook.dto.UploadPhotoDto;
 import com.example.photobook.entity.Photo;
+import com.example.photobook.exception.ResourceForbiddenException;
 import com.example.photobook.helper.PhotoRepositoryHelper;
 import com.example.photobook.mapperToEntity.UploadPhotoDtoMapper;
+import com.example.photobook.repository.AlbumRepository;
 import com.example.photobook.repository.PhotoRepository;
+import com.example.photobook.security.UserContext;
 import com.example.photobook.service.PhotoService;
 import com.example.photobook.util.DownloadingStatusHelper;
 import com.example.photobook.util.LoadingPhotoByURLHelper;
@@ -35,6 +38,8 @@ public class PhotoServiceImpl implements PhotoService {
     private final PhotoRepositoryHelper photoRepositoryHelper;
     private final DownloadingStatusHelper downloadingStatusHelper;
     private final LoadingPhotoByURLHelper loadingPhotoByURLHelper;
+    private final UserContext userContext;
+    private final AlbumRepository albumRepository;
     private final String pathToFiles;
 
     public PhotoServiceImpl(PhotoRepository photoRepository,
@@ -44,6 +49,8 @@ public class PhotoServiceImpl implements PhotoService {
                             PhotoRepositoryHelper photoRepositoryHelper,
                             DownloadingStatusHelper downloadingStatusHelper,
                             LoadingPhotoByURLHelper loadingPhotoByURLHelper,
+                            UserContext userContext,
+                            AlbumRepository albumRepository,
                             @Value("${photobookapp.downloading-directory}") String pathToFiles) {
         this.photoRepository = photoRepository;
         this.modelMapper = modelMapper;
@@ -52,6 +59,8 @@ public class PhotoServiceImpl implements PhotoService {
         this.photoRepositoryHelper = photoRepositoryHelper;
         this.downloadingStatusHelper = downloadingStatusHelper;
         this.loadingPhotoByURLHelper = loadingPhotoByURLHelper;
+        this.userContext = userContext;
+        this.albumRepository = albumRepository;
         this.pathToFiles = pathToFiles;
     }
 
@@ -59,6 +68,11 @@ public class PhotoServiceImpl implements PhotoService {
     public void deletePhoto(Long photoId) {
         Optional<Photo> photoOptional = photoRepository.findById(photoId);
         if (photoOptional.isPresent()) {
+            Photo photo = photoOptional.get();
+            if (!photo.getAlbum().getUserOwner().getUsername()
+                    .equals(userContext.getAuthentication().getName())) {
+                throw new ResourceForbiddenException("You are not owner of this photo");
+            }
             String pathToPhoto = pathToFiles + File.separator + photoOptional.get().getPhotoName();
             photoRepository.deleteById(photoId);
             new File(pathToPhoto).delete();
@@ -80,6 +94,7 @@ public class PhotoServiceImpl implements PhotoService {
     @Override
     public PhotoDto uploadPhotoFromComputer(UploadPhotoDto uploadPhotoDto,
                                             MultipartFile file) throws IOException {
+        checkUserRights(uploadPhotoDto.getAlbumId());
         uploadPhotoDtoValidator.checkPhotoUploadingFromComputer(uploadPhotoDto, file);
         uploadPhotoDto.setPhotoName(buildPhotoName(uploadPhotoDto, file));
         savePhotoOnServer(uploadPhotoDto, file);
@@ -89,6 +104,7 @@ public class PhotoServiceImpl implements PhotoService {
 
     @Override
     public PhotoDto uploadPhotoByUrl(UploadPhotoDto uploadPhotoDto) {
+        checkUserRights(uploadPhotoDto.getAlbumId());
         uploadPhotoDtoValidator.checkPhotoUploadingByUrl(uploadPhotoDto);
         uploadPhotoDto.setPhotoName(buildPhotoName(uploadPhotoDto));
         Photo savedPhoto = photoRepository.save(uploadPhotoDtoMapper.toEntity(uploadPhotoDto));
@@ -102,6 +118,13 @@ public class PhotoServiceImpl implements PhotoService {
         stream.write(bytes);
         stream.flush();
         stream.close();
+    }
+
+    private void checkUserRights(Long albumId) {
+        if (!albumRepository.existsAlbumByIdAndUserOwnerUsername(albumId,
+                userContext.getAuthentication().getName())) {
+            throw new ResourceForbiddenException("You are not the owner of this album");
+        }
     }
 
 }
